requirements.txt
anthropic>=0.39.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit>=1.38.0
PyMuPDF>=1.24.0
python-docx>=1.1.0
tiktoken>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0
